requirements.txt
requests>=2.31.0
openai>=1.3.0
crewai[tools]>=0.11.0
python-dotenv>=1.0.0
PyGithub>=2.1.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1
SQLAlchemy>=2.0.23
aiosqlite>=0.19.0
loguru>=0.7.2
tqdm>=4.66.1
